+++ src/packageurl/contrib/purl2url.py
@@ -186,6 +186,19 @@ def build_rubygems_repo_url(purl):
         return f"https://rubygems.org/gems/{name}"
 
 
+@repo_router.route("pkg:cran/.*")
+def build_cran_repo_url(purl):
+    """
+    Return a cran repo URL from the `purl` string.
+    """
+    purl_data = PackageURL.from_string(purl)
+
+    name = purl_data.name
+    version = purl_data.version
+
+    return f"https://cran.r-project.org/src/contrib/{name}_{version}.tar.gz"
+
+
 @repo_router.route("pkg:npm/.*")
 def build_npm_repo_url(purl):
     """
@@ -223,6 +236,23 @@ def build_pypi_repo_url(purl):
         return f"https://pypi.org/project/{name}/{version}/"
     elif name:
         return f"https://pypi.org/project/{name}/"
+
+
+@repo_router.route("pkg:composer/.*")
+def build_composer_repo_url(purl):
+    """
+    Return a composer repo URL from the `purl` string.
+    """
+    purl_data = PackageURL.from_string(purl)
+
+    name = purl_data.name
+    version = purl_data.version
+    namespace = purl_data.namespace
+
+    if name and version:
+        return f"https://packagist.org/packages/{namespace}/{name}#{version}"
+    elif name:
+        return f"https://packagist.org/packages/{namespace}/{name}"
 
 
 @repo_router.route("pkg:nuget/.*")