+++ src/packageurl/utils.py
@@ -30,12 +30,24 @@ from packageurl import PackageURL
 def get_golang_purl(go_package: str):
     """
     Return a PackageURL object given an imported ``go_package``
+    or go module "name version" string as seen in a go.mod file.
+    >>> get_golang_purl(go_package="github.com/gorilla/mux v1.8.1")
+    PackageURL(type='golang', namespace='github.com/gorilla', name='mux', version='v1.8.1', qualifiers={}, subpath=None)
     """
     if not go_package:
         return
+    version = None
+    # Go package in *.mod files is represented like this
+    # package version
+    # github.com/gorilla/mux v1.8.1
+    # https://github.com/moby/moby/blob/6c10086976d07d4746e03dcfd188972a2f07e1c9/vendor.mod#L51
+    if "@" in go_package:
+        raise Exception(f"{go_package} should not contain ``@``")
+    if " " in go_package:
+        go_package, _, version = go_package.rpartition(" ")
     parts = go_package.split("/")
     if not parts:
         return
     name = parts[-1]
     namespace = "/".join(parts[:-1])
-    return PackageURL(type="golang", namespace=namespace, name=name)
+    return PackageURL(type="golang", namespace=namespace, name=name, version=version)